scanners/xss/requirements.txt
pika==1.3.2
psycopg2-binary==2.9.9
requests==2.31.0
beautifulsoup4==4.12.3
sentinel-core @ file:///app/scanner-core
